move ids to HIR
